moved details tests out to separate file
